Updated database and sql installation scripts
This commit removes old sql installation scripts and organizes everything into a logical structure.

Additionally, for all tables to be added at install time, we are now using GZip for storage in the repository to save space and not require the use of LFS.

The database sql files have also been updated so that they now match the latest eve-server core updates in 0.8.0.
Update SQL table data

This commit updates SQL table data with new schema changes made in 0.8.0